Item13989: Allow selected topics to get guest sessions.
This adds a new context:  sessionRequired,  which is set for scripts and
topics that can't be used without a CGI session, for ex. pages that
submit forms using StrikeOne validation, like UserRegistration.

This also makes a minor change in the order that objects are created
during Foswiki initialization.  The Foswiki->new() session object is
moved down past parsing of the query path, so that the topic is
determined before the session is created.

Needed to change the RegisterTests,  as several of them ran "register"
and "reset password" from the WebHome topic, which would not normally be
done.  Alcso ManageDotPm and CommentPlugin tests need sessions.

Also disable the override for GuestSessions in FoswikiTestCase,  with
these changes, tests should run with sessions disabled.

+++ core/lib/Foswiki/LoginManager.pm
@@ -567,11 +567,13 @@ sub loadSession {
     # Call to getLoggedIn inserts the auth user into the cgi session
     $this->userLoggedIn($authUser)
       unless ( $authUser eq $Foswiki::cfg{DefaultUserLogin}
-        && !$guestSessions );
+        && !$guestSessions
+        && !$session->inContext('sessionRequired') );
 
     # Cleanup unused guest sessions
     if (   $this->{_cgisession}
         && !$guestSessions
+        && !$session->inContext('sessionRequired')
         && $authUser eq $Foswiki::cfg{DefaultUserLogin} )
     {
         $this->{_cgisession}->delete();
@@ -896,7 +898,8 @@ sub userLoggedIn {
 
             # Don't make a session for the guest user.
             unless ( $authUser eq $Foswiki::cfg{DefaultUserLogin}
-                && !$guestSessions )
+                && !$guestSessions
+                && !$session->inContext('sessionRequired') )
 
             {
                 $this->{_cgisession} =
